fix(llma): deduplicate label helpers, add save loading state
Move getLinkedEvaluationLabel and getLinkedSourceLabel to the logic
file where they semantically belong. Reuse getLinkedEvaluationLabel
in the saveToLinkedEvaluation listener instead of duplicating the
logic. Add saving reducer to prevent double-clicks during save.
